refactor: one binding TU; rolling entry points generated per statistic
BindRolling<T>(module, class, name, kernel) registers the two free functions
and four GroupedArray methods a rolling statistic needs, deriving update,
seasonal and seasonal-update from the transform through the combinators in
rolling.h; BindRollingQuantile does the same with the extra p. The 20 named
aliases in rolling.h, the 10 wrappers in rolling.cpp and the 20 GroupedArray
rolling methods go away. Everything binds from src/coreforecast.cpp so the
binders can see the (hidden-visibility) GroupedArray class.

The private C++ free functions rolling_quantile / seasonal_rolling_quantile
now take p right after data, the order the public Python functions and the
GroupedArray methods already used; rolling.py passes it accordingly.
