--- src/diff.cpp
@@ -1,41 +1,0 @@
-#include "bindings.h"
-#include "seasonal.h"
-
-#include "diff.h"
-
-template <typename T> int NumDiffs(const py::array_t<T> data, int max_d) {
-  T out;
-  const auto x = AsContiguous(data);
-  diff::NumDiffs(View(x), std::span<T>{&out, 1}, max_d);
-  return static_cast<int>(out);
-}
-
-template <typename T>
-int NumSeasDiffs(const py::array_t<T> data, int period, int max_d) {
-  RequireNonNegative("season_length", period);
-  T out;
-  const auto x = AsContiguous(data);
-  diff::NumSeasDiffs(View(x), std::span<T>{&out, 1}, period, max_d);
-  return static_cast<int>(out);
-}
-
-template <typename T>
-py::array_t<T> Difference(const py::array_t<T> data, int d) {
-  RequireNonNegative("d", d);
-  const auto x = AsContiguous(data);
-  py::array_t<T> out(x.size());
-  seasonal::Difference(View(x), MutableView(out), d);
-  return out;
-}
-
-template <typename T> void init_diffs_fns(py::module_ &m) {
-  m.def("num_diffs", &NumDiffs<T>);
-  m.def("num_seas_diffs", &NumSeasDiffs<T>);
-  m.def("diff", &Difference<T>);
-}
-
-void init_diffs(py::module_ &m) {
-  py::module_ diffs = m.def_submodule("differences");
-  init_diffs_fns<double>(diffs);
-  init_diffs_fns<float>(diffs);
-}
