+++ tests/cpp/test_stats.cpp
@@ -37,8 +37,7 @@ TEST_CASE_TEMPLATE("Difference and InvertDifference round-trip", T, float,
     }
     // the inverse takes the differenced tail and the d values that preceded it
     std::vector<T> restored(n - d);
-    diff::InvertDifference(In(diffs).subspan(d), In(x).first(d),
-                           Out(restored));
+    diff::InvertDifference(In(diffs).subspan(d), In(x).first(d), Out(restored));
     CheckClose(restored, std::vector<T>(x.begin() + d, x.end()));
   }
   // d = 0 is a copy; d > n is all NaN
@@ -53,27 +52,27 @@ TEST_CASE("KPSS matches statsmodels") {
   // np.cumsum(np.random.default_rng(42).normal(size=64)); the expected values
   // are statsmodels.tsa.stattools.kpss(x, regression="c", nlags=k)[0]
   const std::vector<double> x = {
-      0.30471707975443135,   -0.7352670264860641,  0.015184169320393126,
-      0.955748885711607,     -0.9952863029422294,  -2.2974658098045477,
-      -2.1696254066372624,   -2.4858679989808445,  -2.5026691564851333,
-      -3.3557130840587135,   -2.476315109195885,   -1.6985231737669366,
-      -1.6324924762057207,   -0.5052512692376878,  -0.037741926985642216,
-      -0.8970343898688804,   -0.5282836057863816,  -1.4871662066153806,
-      -0.608715905308108,    -0.6586418162943609,  -0.8435041798396214,
-      -1.5244337242435628,   -0.3018923855695326,  -0.4564218676383347,
-      -0.884749689801442,    -1.2368832402896714,  -0.7045740547363227,
-      -0.3391299903722444,   0.073602621223744,    0.5044236242316267,
-      2.646071225102088,     2.2396562087174727,   1.7274134796459353,
-      0.9136407513980576,    1.5296201739735533,   2.658592466694445,
-      2.54464500903957,      1.704488532077042,    0.8800073163858024,
-      1.5306001042105035,    2.273854275413946,    2.817008543719141,
-      2.1514988364304464,    2.383660159497166,    2.500345968637894,
-      2.7190345653669072,    3.590463343315097,    3.814058892089779,
-      4.492972455161674,     4.560551524650565,    4.8496709233405495,
-      5.48095914917909,      4.023803329323424,    3.704132112966122,
-      3.233759458673327,     2.5948816104299848,   2.319739359203301,
-      3.814680670437697,     2.948849554744454,    3.917127909335935,
-      2.23425813772013,      1.8993731077343552,   2.0621261728393607,
+      0.30471707975443135, -0.7352670264860641, 0.015184169320393126,
+      0.955748885711607,   -0.9952863029422294, -2.2974658098045477,
+      -2.1696254066372624, -2.4858679989808445, -2.5026691564851333,
+      -3.3557130840587135, -2.476315109195885,  -1.6985231737669366,
+      -1.6324924762057207, -0.5052512692376878, -0.037741926985642216,
+      -0.8970343898688804, -0.5282836057863816, -1.4871662066153806,
+      -0.608715905308108,  -0.6586418162943609, -0.8435041798396214,
+      -1.5244337242435628, -0.3018923855695326, -0.4564218676383347,
+      -0.884749689801442,  -1.2368832402896714, -0.7045740547363227,
+      -0.3391299903722444, 0.073602621223744,   0.5044236242316267,
+      2.646071225102088,   2.2396562087174727,  1.7274134796459353,
+      0.9136407513980576,  1.5296201739735533,  2.658592466694445,
+      2.54464500903957,    1.704488532077042,   0.8800073163858024,
+      1.5306001042105035,  2.273854275413946,   2.817008543719141,
+      2.1514988364304464,  2.383660159497166,   2.500345968637894,
+      2.7190345653669072,  3.590463343315097,   3.814058892089779,
+      4.492972455161674,   4.560551524650565,   4.8496709233405495,
+      5.48095914917909,    4.023803329323424,   3.704132112966122,
+      3.233759458673327,   2.5948816104299848,  2.319739359203301,
+      3.814680670437697,   2.948849554744454,   3.917127909335935,
+      2.23425813772013,    1.8993731077343552,  2.0621261728393607,
       2.6483485041986388,
   };
   CHECK(KPSS(In(x), 0) == doctest::Approx(5.063007444851195).epsilon(1e-9));
@@ -121,7 +120,8 @@ TEST_CASE_TEMPLATE("BoxCox transforms round-trip", T, float, double) {
       CAPTURE(x);
       const T y = scalers::BoxCoxTransform<T>(x, lambda, T{0});
       const T back = scalers::BoxCoxInverseTransform<T>(y, lambda, T{0});
-      CHECK(back == doctest::Approx(x).epsilon(std::is_same_v<T, float> ? 1e-4 : 1e-9));
+      CHECK(back ==
+            doctest::Approx(x).epsilon(std::is_same_v<T, float> ? 1e-4 : 1e-9));
     }
   }
   // negative input with a negative lambda has no real transform
